GEODE-1505: merged 2 classes. moved the tests from incorrectly spelled test class (#6757)
* merged 2 classes. moved the tests from incorrectly spelled test class
* ran ./gradlew spotlessApply per review comments on PR
* fixed the test names per review comments by Donal Evans
* renamed test cases to more appropriate names as per review comments received
